requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
PyJWT==2.3.0
python-dotenv==0.19.0 